Made classes, splitted files, updated readme and depedencies, added inversion of control
