+++ lib/index.d.ts
@@ -1,1 +1,2 @@
-export {};
+import { Tscov } from './tscov/tscov';
+export declare function index(): Tscov;
