--- lib/tscov.d.ts
@@ -1,2 +1,0 @@
-import ts from 'typescript';
-export declare function lint(project: string, detail: boolean, debug: boolean, files?: any, oldProgram?: ts.Program): Promise<any>;
