+++ lib/tscov/options.d.ts
@@ -1,0 +1,4 @@
+export declare class Options {
+    constructor();
+    showOptions(): void;
+}
