+++ lib/tscov/types/json-config.type.d.ts
@@ -1,0 +1,9 @@
+export declare type JsonConfig = {
+    extends?: string;
+    compilerOptions?: {
+        [name: string]: any;
+    };
+    include?: string[];
+    exclude?: string[];
+    files?: string[];
+};
