+++ package.json
@@ -1,19 +1,18 @@
 {
     "name": "tscov",
-    "version": "1.2.1",
+    "version": "1.3.0",
     "description": "TypeScript CLI to calculate type coverage",
-    "main": "lib/core.js",
-    "types": "lib/core.d.ts",
+    "main": "lib/index.js",
+    "types": "lib/index.d.ts",
     "repository": "https://github.com/jeroenouw/tscov.git",
     "author": "Jeroen Ouwehand",
     "license": "MIT",
     "scripts": {
-        "start": "nodemon --watch 'src/**/*.ts' --exec 'ts-node' src/index.ts",
-        "create": "npm run build && npm run test",
+        "start": "npm run build && npm run global",
         "build": " rm -rf ./lib && tsc -p .",
-        "test": "sudo npm i -g && tscov",
+        "global": "sudo npm i -g && tscov",
         "link-upstream": "git remote add upstream https://github.com/jeroenouw/tscov",
-        "sync": "git fetch && git checkout master && git merge upstream/master && git push",
+        "sync": "git fetch origin && git checkout master && git pull && git checkout <your-branch> && git merge master && git push",
         "reset": "rm -rf ./node_modules ./package-lock.json ./lib && npm install"
     },
     "typeCoverage": {
@@ -26,14 +25,19 @@
         "cli",
         "tscov",
         "typescript",
+        "javascript",
+        "type",
+        "coverage",
         "nodejs"
     ],
     "dependencies": {
         "commander": "^2.19.0",
         "figlet": "^1.2.1",
         "glob": "7.1.3",
+        "inversify": "^5.0.1",
         "minimist": "1.2.0",
         "prompts": "^2.0.1",
+        "reflect-metadata": "^0.1.13",
         "tslib": "1.9.3"
     },
     "devDependencies": {
@@ -41,16 +45,13 @@
         "@types/jasmine": "3.3.7",
         "@types/minimist": "1.2.0",
         "@types/node": "10.12.18",
-        "clean-release": "2.7.0",
-        "clean-scripts": "1.9.2",
         "jasmine": "3.3.1",
         "no-unused-export": "1.7.0",
-        "nodemon": "^1.18.9",
         "standard": "12.0.1",
         "ts-node": "^7.0.1",
         "tslint": "5.12.1",
         "tslint-config-standard": "8.0.1",
         "tslint-sonarts": "1.8.0",
-        "typescript": "^3.2.4"
+        "typescript": "^3.3.1"
     }
 }
