+++ src/tscov/types/json-config.type.ts
@@ -1,0 +1,8 @@
+export type JsonConfig = {
+    extends?: string,
+    compilerOptions?: { [name: string]: any },
+    include?: string[],
+    exclude?: string[],
+    files?: string[]
+}
+
